+++ src/class/log.ts
@@ -1,4 +1,5 @@
 import * as vs from 'vscode';
+import cp from 'child_process';
 
 export class Log {
     constructor() {
@@ -10,11 +11,17 @@ export class Log {
     e(message: string) { this.channel.appendLine(`${this.now()} E: ${message}`); }
     vsI(message: string) {
         this.i(message);
-        vs.window.showInformationMessage(message);
+        vs.window.showInformationMessage(`TurboGerrit: ${message}`);
     }
     vsE(message: string) {
         this.e(message);
-        vs.window.showErrorMessage(message); 
+        vs.window.showErrorMessage(`TurboGerrit: ${message}`);
+    }
+
+    logCpContent(err: cp.ExecException | null, stdout: string, stderr: string) {
+        if (stdout !== undefined && stdout.length > 0) this.e(`STDOUT: ${stdout}`);
+        if (err !== undefined && err !== null) this.e(`ERR CODE: ${err.code}`);
+        if (err !== undefined && err !== null) this.e(`ERR: ${err.message}`);
     }
 
     private now(): string {