+++ CHANGELOG.md
@@ -1,4 +1,12 @@
 # Change Log
+## [0.0.2]
+
+- using plain ssh instead git-review
+- workspace config added for:
+   1. reviewers, 
+   2. username
+   3. custom .gitreview path
+
 ## [0.0.1]
 
 - Initial release. Push to Gerrit available as command and from vcs context menu.